requirements.txt
python-gitlab>=3.0.0
requests>=2.25.1
python-dateutil>=2.8.2